RFC: Modify is_sos_vm to use vm_id
V2:
In many functions, acrn_vm is being directly used.
This patch is an example patch, to get comments from
reviewers, whether the proposed design choice is a
good idea.

This patch uses vm_type to check whether the vm is SOS.
This way, instead of requiring to pass the whole acrn_vm,
we can just use vm_type to determine same.

This way we may be able to forego lot of forward declarations
of struct acrn_vm in unrelated files, and reduce the
number of MISRA-C violations.

The patch also replaces usage of of acrn_vm, where
the pointers are being compared, by using the vm_id.
The vm_id is being stored in ptirq_remapping_info.

V1:
Was wrongly using vm_id to determine whether vm is SOS.

Signed-off-by: Arindam Roy <arindam.roy@intel.com>